Refactor propTypes in AnchorForCommentsOnly component
Those types were duplicated!
